Adapt code to test MIT mode
