+++ stm32/h7a3zi-q/Core/Inc/uartRingBufDMA.h
@@ -8,6 +8,10 @@
 #ifndef INC_UARTRINGBUFDMA_H_
 #define INC_UARTRINGBUFDMA_H_
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 #include "main.h"
 
 #include <stdint.h>
@@ -22,5 +26,8 @@ void Ringbuf_Init (void);
  * */
 void Ringbuf_Reset (UART_HandleTypeDef *huart);
 
+#ifdef __cplusplus
+}
+#endif
 
 #endif /* INC_UARTRINGBUFDMA_H_ */
