+++ stm32/h7a3zi-q/Periph/Inc/comUtils_UART2.h
@@ -2,6 +2,10 @@
 #ifndef COMUTILS_UART2_H
 #define COMUTILS_UART2_H
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 #include <stdint.h>
 
 /* Search for a word in a wordlist inside the buffer and deletes unfound words
@@ -10,5 +14,9 @@
  */
 
 char* searchWord(char* buffertocopyfrom);
+
+#ifdef __cplusplus
+}
+#endif
 
 #endif /* CONMUTILS_UART2_H */
