+++ stm32/h7a3zi-q/Periph/Src/CanMotorServo.c
@@ -26,26 +26,26 @@ void CanMotorServo_Init()
     }
 }
 
-void HAL_FDCAN_RxFifo0Callback(FDCAN_HandleTypeDef* hfdcan, uint32_t RxFifo0ITs)
-{
-    if ((RxFifo0ITs & FDCAN_IT_RX_FIFO0_NEW_MESSAGE) != RESET)
-    {
-        /* Retreive Rx messages from RX FIFO0 */
-        if (HAL_FDCAN_GetRxMessage(hfdcan, FDCAN_RX_FIFO0, &RxHeader, RxData) !=
-            HAL_OK)
-        {
-            /* Reception Error */
-            Error_Handler();
-        }
-
-        if (HAL_FDCAN_ActivateNotification(
-                hfdcan, FDCAN_IT_RX_FIFO0_NEW_MESSAGE, 0) != HAL_OK)
-        {
-            /* Notification Error */
-            Error_Handler();
-        }
-    }
-}
+//void HAL_FDCAN_RxFifo0Callback(FDCAN_HandleTypeDef* hfdcan, uint32_t RxFifo0ITs)
+//{
+//    if ((RxFifo0ITs & FDCAN_IT_RX_FIFO0_NEW_MESSAGE) != RESET)
+//    {
+//        /* Retreive Rx messages from RX FIFO0 */
+//        if (HAL_FDCAN_GetRxMessage(hfdcan, FDCAN_RX_FIFO0, &RxHeader, RxData) !=
+//            HAL_OK)
+//        {
+//            /* Reception Error */
+//            Error_Handler();
+//        }
+//
+//        if (HAL_FDCAN_ActivateNotification(
+//                hfdcan, FDCAN_IT_RX_FIFO0_NEW_MESSAGE, 0) != HAL_OK)
+//        {
+//            /* Notification Error */
+//            Error_Handler();
+//        }
+//    }
+//}
 
 void CanMotorServo_Transmit(uint32_t id, const uint8_t* data, uint32_t bytes)
 {
